removed includes from headers
corrected location of some protos
removed un-needed includes from source files
