+++ inc/chksum64.h
@@ -1,4 +1,7 @@
-
+//
+// Copyright (c) 2017 The Altra64 project contributors
+// See LICENSE file in the project root for full license information.
+//
 
 #ifndef _CHKSUM64_H
 #define	_CHKSUM64_H
