+++ src/chksum64.c
@@ -18,8 +18,7 @@
     along with this program; if not, write to the Free Software
     Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA
 */
-#include <stdio.h>
-#include <stdlib.h>
+
 #include <libdragon.h>
 
 #include "sys.h"
