+++ src/everdrive.c
@@ -4,14 +4,9 @@
 // See LICENSE file in the project root for full license information.
 //
 
-#include "types.h"
 #include "everdrive.h"
-#include <libdragon.h>
-#include <stdio.h>
 #include "sys.h"
 #include "errors.h"
-//#include "rom.h"
-#include "disk.h"
 
 #define CMD0  0x40    // software reset
 #define CMD1  0x41    // brings card out of idle state
