+++ src/fat.c
@@ -4,17 +4,15 @@
 // See LICENSE file in the project root for full license information.
 //
 
-#include <console.h>
 #include <malloc.h>
+#include <libdragon.h>
 
 #include "fat.h"
 #include "disk.h"
 #include "mem.h"
 #include "everdrive.h"
-#include "sys.h"
-#include <libdragon.h>
 #include "strlib.h"
-
+#include "errors.h"
