+++ src/mp3.c
@@ -7,7 +7,6 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
-#include <stdint.h>
 #include <mad.h>
 #include "fat.h"
 #include "mp3.h"
